add format in news for search request fixed
