+++ src/index.js
@@ -1,7 +1,6 @@
 import './js/mobile_menu';
 import './js/categories/init';
 import './js/calendar/index';
-import './js/categories/myPagination'; //current version for testing!
 import './js/pagination.js';
 import './js/homepage-render';
 import './js/fetch_weather';
